Extract interface out of Paper and rename implementation
To support future changes the Paper must be polymorphic.
The current implementation is used in tests to assert
that specific behaviours were triggered.
The actions taken are:
  - Extract an interface which is named 'Paper'
  - Rename current implementation to 'TestPaper'
  - Push all methods into the interface which do not assert behaviour

Issue #9
